refactor: show logs to each application and database

+++ apps/dokploy/server/wss/docker-container-terminal.ts
@@ -30,6 +30,7 @@ export const setupDockerContainerTerminalWebSocketServer = (
 		const url = new URL(req.url || "", `http://${req.headers.host}`);
 		const containerId = url.searchParams.get("containerId");
 		const activeWay = url.searchParams.get("activeWay");
+		const serverId = url.searchParams.get("serverId");
 		const { user, session } = await validateWebSocketRequest(req);
 
 		if (!containerId) {
@@ -42,41 +43,45 @@ export const setupDockerContainerTerminalWebSocketServer = (
 			return;
 		}
 		try {
-			const shell = getShell();
-			const ptyProcess = spawn(
-				shell,
-				["-c", `docker exec -it ${containerId} ${activeWay}`],
-				{
-					name: "xterm-256color",
-					cwd: process.env.HOME,
-					env: process.env,
-					encoding: "utf8",
-					cols: 80,
-					rows: 30,
-				},
-			);
+			if (serverId) {
+				// const server = await findServerById(serverId);
+			} else {
+				const shell = getShell();
+				const ptyProcess = spawn(
+					shell,
+					["-c", `docker exec -it ${containerId} ${activeWay}`],
+					{
+						name: "xterm-256color",
+						cwd: process.env.HOME,
+						env: process.env,
+						encoding: "utf8",
+						cols: 80,
+						rows: 30,
+					},
+				);
 
-			ptyProcess.onData((data) => {
-				ws.send(data);
-			});
-			ws.on("close", () => {
-				ptyProcess.kill();
-			});
-			ws.on("message", (message) => {
-				try {
-					let command: string | Buffer[] | Buffer | ArrayBuffer;
-					if (Buffer.isBuffer(message)) {
-						command = message.toString("utf8");
-					} else {
-						command = message;
+				ptyProcess.onData((data) => {
+					ws.send(data);
+				});
+				ws.on("close", () => {
+					ptyProcess.kill();
+				});
+				ws.on("message", (message) => {
+					try {
+						let command: string | Buffer[] | Buffer | ArrayBuffer;
+						if (Buffer.isBuffer(message)) {
+							command = message.toString("utf8");
+						} else {
+							command = message;
+						}
+						ptyProcess.write(command.toString());
+					} catch (error) {
+						// @ts-ignore
+						const errorMessage = error?.message as unknown as string;
+						ws.send(errorMessage);
 					}
-					ptyProcess.write(command.toString());
-				} catch (error) {
-					// @ts-ignore
-					const errorMessage = error?.message as unknown as string;
-					ws.send(errorMessage);
-				}
-			});
+				});
+			}
 		} catch (error) {
 			// @ts-ignore
 			const errorMessage = error?.message as unknown as string;
